COMMERCE-12879 fix WebKeys + navigation category
