+++ modules/apps/commerce/commerce-shipping-engine/src/main/java/com/liferay/commerce/shipping/engine/internal/constants/FunctionCommerceShippingEngineWebKeys.java
@@ -10,12 +10,6 @@ package com.liferay.commerce.shipping.engine.internal.constants;
  */
 public class FunctionCommerceShippingEngineWebKeys {
 
-	public static final String COMMERCE_SHIPPING_METHOD_ENGINE_KEY =
-		"COMMERCE_SHIPPING_METHOD_ENGINE_KEY";
-
-	public static final String COMMERCE_SHIPPING_METHOD_ID =
-		"COMMERCE_SHIPPING_METHOD_ID";
-
 	public static final String IS_DEFAULT_VALUE = "IS_DEFAULT_VALUE";
 
 	public static final String SHIPPING_METHOD_TYPE_SETTINGS =
